Remove chance for GoblinTrading fallback trades + small zendoc fixes

+++ src/main/java/com/smokeythebandicoot/witcherycompanion/integrations/crafttweaker/ExtendedPlayerHandler.java
@@ -330,7 +330,7 @@ public class ExtendedPlayerHandler {
         }
 
         @ZenMethod
-        public double getMaxLevel() {
+        public int getMaxLevel() {
             if (stats == null || stats.getCreatureTraitType() == null) {
                 return -1;
             }
